Change View::GetIndexOf() to return absl::optional<size_t>.
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3781145
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1027897}

+++ ui/views/controls/menu/menu_model_adapter_unittest.cc
@@ -231,7 +231,7 @@ void CheckSubmenu(const RootModel& model,
       continue;
     }
     // Check placement.
-    EXPECT_EQ(i, static_cast<size_t>(submenu->GetSubmenu()->GetIndexOf(item)));
+    EXPECT_EQ(i, submenu->GetSubmenu()->GetIndexOf(item));
 
     // Check type.
     switch (model_item.type) {
@@ -315,7 +315,7 @@ TEST_F(MenuModelAdapterTest, BasicTest) {
     }
 
     // Check placement.
-    EXPECT_EQ(i, static_cast<size_t>(menu->GetSubmenu()->GetIndexOf(item)));
+    EXPECT_EQ(i, menu->GetSubmenu()->GetIndexOf(item));
 
     // Check type.
     switch (model_item.type) {
